+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Spawn plugin will be documented in this file.
 ## Unreleased
 
 - Add domain renewal abilities (spawn_get_domain_renewal_info, spawn_renew_domain)
+- Create single source of truth for tier configuration
 
 ## [0.2.0] - 2025-02-09